pyproject.toml
[project]
name = "qa-rag-agent"
version = "0.2.0"
description = "Q&A RAG agent with LangGraph, ChromaDB, persistent memory, and observability"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "User", email = "user@example.com" }]

dependencies = [
    # LangChain stack
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    # LangGraph
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    # LLM providers
    "langchain-openai>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-ollama>=0.2.0",
    # Embeddings
    "langchain-huggingface>=0.1.0",
    "sentence-transformers>=2.2.2",
    # Vector store
    "langchain-chroma>=0.2.0",
    "chromadb>=0.5.0",
    # Document loaders
    "pypdf>=3.0.0",
    "unstructured>=0.12.0",
    "markdown>=3.5.0",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
    # UI
    "streamlit>=1.40.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    # V2: Persistent memory
    "sqlalchemy>=2.0.25",
    # V2: Observability
    "langfuse>=2.0.0",
    "langsmith>=0.1.0",
    # V2: Web search
    "tavily-python>=0.4.0",
    # V7: Checkpointing
    "langgraph-checkpoint-sqlite>=2.0.0",
    # API layer
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.7",
    "modal>=1.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "pytest-mock>=3.12.0",
    "graphviz>=0.20.0",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.7",
]
deploy = [
    "modal>=0.72.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py312"
